# YJIT: Local variable register allocation