This PR introduces a register allocator for local variables. This PR contains the following changes for it:

* Migrate `Context::reg_temps` from `u8` to `[Option<RegTemp>; 5]`
    * `RegTemp` is either `RegTemp::Stack(u8)` or `RegTemp::Local(u8)`.
    * The encoder uses 1-5 bits for each of the 5 registers. 1 bit for `Option`, 1 bit for enum variant, and 3 bits for index.
* `Opnd::Stack` can now be assigned to any register while it still defaults to `stack_idx % 5` (register bias)
    * When the register at `stack_idx % 5` is already used, it uses an unused register with the smallest index.
* Allocate a register for a local variable on setivar
    * Convert local variable access from `Opnd::Mem` to `Opnd::Stack`
    * It defaults to the register at `4 - (local_idx % 5)`. If it's taken, it uses an unused register with the largest index.

## Out of Scope

Method arguments are local variables but they aren't set by `setivar`, so it's not supported in this PR. 

I'm thinking of working on that next, but I think the current optimization should be merged separately to keep the patches simple and to allow benchmarking them independently.

## Benchmark

It gives 1-2% speedup to some benchmarks. `railsbench` specifically seems to consistently show a 2% speedup.

```
before: ruby 3.4.0dev (2024-07-09T17:22:29Z master 6f6aff56b1) +YJIT [x86_64-linux]
after: ruby 3.4.0dev (2024-07-12T09:13:23Z yjit-local-regs 1ad043bd3e) +YJIT [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  -------------  ------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
activerecord    302.0        0.2         302.2       0.4         1.00           1.00
chunky-png      487.5        0.1         486.1       0.2         0.99           1.00
erubi-rails     1275.5       0.1         1259.9      0.1         1.01           1.01
hexapdf         1876.7       2.5         1869.0      1.3         0.98           1.00
liquid-c        76.5         0.7         75.4        0.5         0.96           1.01
liquid-compile  75.7         0.5         74.7        0.9         0.90           1.01
liquid-render   98.8         0.3         97.8        0.5         0.96           1.01
lobsters        985.5        1.6         979.9       1.4         0.97           1.01
mail            155.9        0.4         155.6       0.5         0.95           1.00
psych-load      1803.7       0.1         1826.4      0.1         0.99           0.99
railsbench      2505.3       0.7         2457.8      0.8         0.99           1.02
rubocop         130.5        6.6         131.1       7.1         0.94           1.00
ruby-lsp        145.3        0.9         143.6       1.0         0.93           1.01
sequel          92.6         0.6         92.7        0.4         0.99           1.00
--------------  -----------  ----------  ----------  ----------  -------------  ------------
Legend:
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.

Output:
./data/output_1343.csv
```

## Memory Usage

The current master always uses 8 bits for a `reg_temps`. This PR uses 5-25 bits. So, on average, `bytes_per_context` is increased. However, it doesn't seem to have a significant impact on overall `yjit_alloc_size`, thanks to context caching.

With 25 itrs on lobsters, 

### Before

```
yjit_alloc_size:          27,728,430
context_data_bytes:          341,320
context_cache_bytes:          16,384
num_contexts_encoded:        192,683
bytes_per_context:              1.77
context_cache_hit_rate:      141,224 (73.3%)
```

### After

```
yjit_alloc_size:          27,608,015
context_data_bytes:          390,787
context_cache_bytes:          20,480
num_contexts_encoded:        191,803
bytes_per_context:              2.04
context_cache_hit_rate:      136,528 (71.2%)
```