@maximecb 2024-07-12
Hi Kokubun!

I took a first look through this PR just now. Props for getting the changes passing all the tests 🙏 

I think your encoding is pretty good in terms of compactness 👍 Happy that it doesn't increase YJIT alloc size significantly.

Can you report on the inline/outlined code size before and after?

I'm a bit surprised that we don't get more of a speedup on benchmarks. Do we get much of a speedup on any of the smaller benchmarks? How does protoboeuf look specifically? Also curious if you've inspected the generated code to see that it's not doing anything obviously suboptimal.

I think performance should not ultimately be the criteria by which we judge this change because it seems like we can probably make more iterative improvements to code quality once we have a better unified register allocator, but it would still be good to know how code size is affected and if there are obvious things that could be improved.