@k0kubun 2024-07-12
> Can you report on the inline/outlined code size before and after?

With 25 itrs on lobsters (the same run as the one in the PR description), the code size seems to be reduced.

### Before
```
inline_code_size:          7,561,968
outlined_code_size:        6,312,621
code_region_size:         15,466,496
code_region_overhead:      1,591,907 (10.3%)
```

### After
```
inline_code_size:          7,513,927
outlined_code_size:        6,423,682
code_region_size:         15,433,728
code_region_overhead:      1,496,119 ( 9.7%)
```

> Also curious if you've inspected the generated code to see that it's not doing anything obviously suboptimal.

I have. This is the generated code for `def test; a = 1; a; end; test`:

```asm
  # Block: test@-e:1
  # reg_mappings: [None, None, None, None, None]
  # Insn: 0000 putobject_INT2FIX_1_ (stack_size: 0)
  # reg_mappings: [None, None, None, None, None] -> [Some(Stack(0)), None, None, None, None]
  0x5f84e6dfd058: mov esi, 3
  # Insn: 0001 setlocal_WC_0 (stack_size: 1)
  # reg_mappings: [Some(Stack(0)), None, None, None, None] -> [Some(Stack(0)), None, None, None, Some(Local(0))]
  0x5f84e6dfd05d: mov r10, rsi
  # Insn: 0003 getlocal_WC_0 (stack_size: 0)
  # reg_mappings: [None, None, None, None, Some(Local(0))] -> [Some(Stack(0)), None, None, None, Some(Local(0))]
  0x5f84e6dfd060: mov rsi, r10
```

which is the intended outcome as of this PR, and for `def test; a = 1; puts; a; end; test` with a spill on a C call:

```asm
  # Block: test@-e:1
  # reg_mappings: [None, None, None, None, None]
  # Insn: 0000 putobject_INT2FIX_1_ (stack_size: 0)
  # reg_mappings: [None, None, None, None, None] -> [Some(Stack(0)), None, None, None, None]
  0x5f28bbb48058: mov esi, 3
  # Insn: 0001 setlocal_WC_0 (stack_size: 1)
  # reg_mappings: [Some(Stack(0)), None, None, None, None] -> [Some(Stack(0)), None, None, None, Some(Local(0))]
  0x5f28bbb4805d: mov r10, rsi
  # Insn: 0003 putself (stack_size: 0)
  # reg_mappings: [None, None, None, None, Some(Local(0))] -> [Some(Stack(0)), None, None, None, Some(Local(0))]
  0x5f28bbb48060: mov rsi, qword ptr [r13 + 0x18]

  # Insn: 0004 opt_send_without_block (stack_size: 1)
  # call to Object#puts
  ...
  # call C function
  # spill_temps: [Some(Stack(0)), None, None, None, Some(Local(0))] -> [None, None, None, None, None]
  0x5f28bbb480f9: mov qword ptr [rbx], rsi
  0x5f28bbb480fc: mov qword ptr [rbx - 0x20], r10
  ...

  # reg_mappings: [Some(Stack(0)), None, None, None, None]
  # Insn: 0006 pop (stack_size: 1)
  # Insn: 0007 getlocal_WC_0 (stack_size: 0)
  # reg_mappings: [None, None, None, None, None] -> [Some(Stack(0)), None, None, None, None]
  0x5f28bbb4811d: mov rsi, qword ptr [rbx - 0x20]
```

which is also what I intended.

> Do we get much of a speedup on any of the smaller benchmarks? How does protoboeuf look specifically?

Not much on smaller benchmarks either. It doesn't seem to speed up protoboeuf 🤔 

```
before: ruby 3.4.0dev (2024-07-09T17:22:29Z master 6f6aff56b1) +YJIT [x86_64-linux]
after: ruby 3.4.0dev (2024-07-12T09:13:23Z yjit-local-regs 1ad043bd3e) +YJIT [x86_64-linux]

----------  -----------  ----------  ----------  ----------  -------------  ------------
bench       before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
protoboeuf  54.2         0.8         54.4        0.9         0.94           1.00
----------  -----------  ----------  ----------  ----------  -------------  ------------
```

I haven't looked into the generated code of protoboeuf, so I can take a look at that and see if anything can be easily improved. Looking at the source code, `buff` and `index` are not locals set by `setlocal` but method arguments, which are not yet optimized in this PR, so it might partly explain why it doesn't have a significant impact.