@maximecb 2024-07-12
It's nice to see a reduction in inline code size even if it's small. The reg mapping is a lot more readable in the disassembly output which is also nice.

> method arguments, which are not yet optimized in this PR, so it might partly explain why it doesn't have a significant impact.

Supporting method arguments would probably have a big impact as they are the most used locals. Would it be difficult to add to this design? Not asking you to do it in this PR.

I won't push the performance question more for this PR. I think we need to do more in terms of improving the quality of generated code to maximize performance, but I assume we can do it as part of followup PRs.