@maximecb 2024-07-12
The PR seems good from a technical standpoint, almost ready to merge.

I wrote a few more comments on the PR. They are all relatively trivial things with respect to naming, to help make the code easier to understand.